+++ tests/unit/hooks.test.ts
@@ -25,6 +25,27 @@ test('call a function after browser launched', async () => {
     .end()
 })
 
+test('call a function before each step', async () => {
+  const fn = jest.fn(function () {
+    expect(Object.getPrototypeOf(this)).toBe(Rize.prototype)
+  })
+  const instance = new Rize({ beforeEachStep: fn })
+  await instance
+    .execute(() => expect(fn).toHaveBeenCalledTimes(1))
+    .end()
+})
+
+test('call a function after each step', async () => {
+  const fn = jest.fn(function () {
+    expect(Object.getPrototypeOf(this)).toBe(Rize.prototype)
+  })
+  const instance = new Rize({ afterEachStep: fn })
+  await instance
+    .execute(() => 0)
+    .execute(() => expect(fn).toHaveBeenCalledTimes(1))
+    .end()
+})
+
 test('call a function before browser exit', async () => {
   const fn = jest.fn()
   const instance = new Rize({ beforeExit: fn })